+++ projects/packages/jetpack-mu-wpcom/src/features/wpcom-media/wpcom-external-media-import.js
@@ -1,1 +1,0 @@
-console.log( 'Hello, Import Media Page' ); // eslint-disable-line no-console