+++ projects/packages/my-jetpack/src/products/class-scan.php
@@ -7,11 +7,9 @@
 
 namespace Automattic\Jetpack\My_Jetpack\Products;
 
-use Automattic\Jetpack\Connection\Client;
 use Automattic\Jetpack\My_Jetpack\Module_Product;
 use Automattic\Jetpack\My_Jetpack\Wpcom_Products;
 use Automattic\Jetpack\Redirect;
-use Jetpack_Options;
 use WP_Error;
 
 /**
@@ -32,6 +30,13 @@ class Scan extends Module_Product {
 	 * @var string
 	 */
 	public static $module_name = 'scan';
+
+	/**
+	 * The category of the product
+	 *
+	 * @var string
+	 */
+	public static $category = 'security';
 
 	/**
 	 * The feature slug that identifies the paid plan
@@ -115,47 +120,6 @@ class Scan extends Module_Product {
 	}
 
 	/**
-	 * Hits the wpcom api to check scan status.
-	 *
-	 * @todo Maybe add caching.
-	 *
-	 * @return Object|WP_Error
-	 */
-	private static function get_state_from_wpcom() {
-		static $status = null;
-
-		if ( $status !== null ) {
-			return $status;
-		}
-
-		$site_id = Jetpack_Options::get_option( 'id' );
-
-		$response = Client::wpcom_json_api_request_as_blog( sprintf( '/sites/%d/scan', $site_id ) . '?force=wpcom', '2', array( 'timeout' => 2 ), null, 'wpcom' );
-
-		if ( 200 !== wp_remote_retrieve_response_code( $response ) ) {
-			$status = new WP_Error( 'scan_state_fetch_failed' );
-			return $status;
-		}
-
-		$body   = wp_remote_retrieve_body( $response );
-		$status = json_decode( $body );
-		return $status;
-	}
-
-	/**
-	 * Checks whether the current plan (or purchases) of the site already supports the product
-	 *
-	 * @return boolean
-	 */
-	public static function has_required_plan() {
-		$scan_data = static::get_state_from_wpcom();
-		if ( is_wp_error( $scan_data ) ) {
-			return false;
-		}
-		return is_object( $scan_data ) && isset( $scan_data->state ) && 'unavailable' !== $scan_data->state;
-	}
-
-	/**
 	 * Checks whether the Product is active
 	 *
 	 * Scan is not actually a module. Activation takes place on WPCOM. So lets consider it active if jetpack is active and has the plan.
@@ -163,7 +127,7 @@ class Scan extends Module_Product {
 	 * @return boolean
 	 */
 	public static function is_active() {
-		return static::is_jetpack_plugin_active() && static::has_required_plan();
+		return static::is_jetpack_plugin_active();
 	}
 
 	/**