+++ projects/packages/publicize/src/class-publicize-utils.php
@@ -109,4 +109,15 @@ class Publicize_Utils {
 			throw new \Exception( esc_html( "Method $method can only be called on WordPress.com." ) );
 		}
 	}
+
+	/**
+	 * Check if the new module endpoint is available in the used Jetpack version.
+	 * We need the module status in response that's why we do the version check https://github.com/Automattic/jetpack/pull/41461/files#diff-f8e5ef1115599de750b64143dd1901554254eddd95ab4371b6b6b3b2a5914224R638-R642.
+	 * More: https://github.com/Automattic/jetpack/pull/41596.
+	 *
+	 * @return bool
+	 */
+	public static function should_use_jetpack_module_endpoint() {
+		return class_exists( 'Jetpack' ) && defined( 'JETPACK__VERSION' ) && ( version_compare( (string) JETPACK__VERSION, '14.3', '>=' ) );
+	}
 }