+++ projects/plugins/boost/app/modules/optimizations/minify/class-minify-js.php
@@ -3,18 +3,20 @@
 namespace Automattic\Jetpack_Boost\Modules\Optimizations\Minify;
 
 use Automattic\Jetpack_Boost\Contracts\Changes_Page_Output;
+use Automattic\Jetpack_Boost\Contracts\Has_Activate;
+use Automattic\Jetpack_Boost\Contracts\Has_Deactivate;
 use Automattic\Jetpack_Boost\Contracts\Optimization;
 use Automattic\Jetpack_Boost\Contracts\Pluggable;
 use Automattic\Jetpack_Boost\Lib\Minify\Concatenate_JS;
 
-class Minify_JS implements Pluggable, Changes_Page_Output, Optimization {
+class Minify_JS implements Pluggable, Changes_Page_Output, Optimization, Has_Activate, Has_Deactivate {
 
 	public static $default_excludes = array( 'jquery', 'jquery-core', 'underscore', 'backbone' );
 
 	public function setup() {
 		require_once JETPACK_BOOST_DIR_PATH . '/app/lib/minify/functions-helpers.php';
 
-		jetpack_boost_minify_setup();
+		jetpack_boost_minify_init();
 
 		if ( jetpack_boost_page_optimize_bail() ) {
 			return;
@@ -46,5 +48,14 @@ class Minify_JS implements Pluggable, Changes_Page_Output, Optimization {
 		// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
 		$wp_scripts                         = new Concatenate_JS( $wp_scripts );
 		$wp_scripts->allow_gzip_compression = true; // @todo - used constant ALLOW_GZIP_COMPRESSION = true if not defined.
+	}
+
+	public static function activate() {
+		jetpack_boost_minify_activation();
+	}
+
+	public static function deactivate() {
+		jetpack_boost_page_optimize_cleanup_cache( 'js' );
+		jetpack_boost_minify_deactivation();
 	}
 }